Implement strategy pattern for validation
All validation is now handled by separate classes that implement the
IValidationStrategy interface. The validation type is stored as a
parameter within the CustomerBase class. This value is by the child
classes when they are instantiated. Then when the Validate() method is
called it will use the appropriate Validate() method from the validation
classes.